requirements.txt
matplotlib
blessings==1.7
gpustat==0.5.0
lmdb==0.94
numpy==1.16.2
nvidia-ml-py3==7.352.0
opencv-python==4.1.0.25
Pillow==5.4.1
psutil==5.6.2
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tensorboardX==1.6
# warpctc-pytorch==0.1
Augmentor==0.2.3